+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - D&D 4.1.x required
 - Wand and Rods now use the Rod and Wand equipment type rather than consumable.
+- Macro activities will auto roll.
 
 # 6.0.31